refactor: StatusCode 추가에 따른 변경
BadRequest 400
Conflict 409
NotFound 404

+++ src/main/java/roomescape/service/ThemeService.java
@@ -4,6 +4,9 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.common.exception.ConflictException;
+import roomescape.common.exception.NotFoundException;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
 import roomescape.domain.Theme;
 import roomescape.domain.vo.Name;
@@ -15,10 +18,11 @@ import roomescape.dto.response.AvailableTimeResponseDto;
 @Transactional(readOnly = true)
 public class ThemeService {
     private final ThemeDao themeDao;
+    private final ReservationDao reservationDao;
 
-
-    public ThemeService(ThemeDao themeDao) {
+    public ThemeService(ThemeDao themeDao, ReservationDao reservationDao) {
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
     }
 
     public List<Theme> findAll() {
@@ -28,14 +32,14 @@ public class ThemeService {
 
     public Theme findById(Long id) {
         return themeDao.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 테마입니다."));
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 테마입니다."));
     }
 
     @Transactional
     public Theme create(ThemeRequestDto themeRequest) {
         Name name = new Name(themeRequest.name());
         if (themeDao.existsByName(name)) {
-            throw new IllegalArgumentException("이미 존재하는 테마 이름입니다.");
+            throw new ConflictException("이미 존재하는 테마 이름입니다.");
         }
 
         Theme theme = new Theme(name, themeRequest.thumbnailUrl(), themeRequest.description());
@@ -45,7 +49,10 @@ public class ThemeService {
     @Transactional
     public void delete(Long id) {
         Theme theme = themeDao.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 테마입니다."));
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 테마입니다."));
+        if (reservationDao.existsByThemeId(theme.getId())) {
+            throw new ConflictException("예약이 존재하여 테마를 삭제할 수 없습니다.");
+        }
 
         themeDao.delete(theme.getId());
     }
