+++ src/main/java/roomescape/service/TimeService.java
@@ -4,6 +4,9 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.common.exception.ConflictException;
+import roomescape.common.exception.NotFoundException;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.TimeDao;
 import roomescape.domain.Time;
 import roomescape.dto.request.TimeRequestDto;
@@ -12,9 +15,11 @@ import roomescape.dto.request.TimeRequestDto;
 @Transactional(readOnly = true)
 public class TimeService {
     private final TimeDao timeDao;
+    private final ReservationDao reservationDao;
 
-    public TimeService(TimeDao timeDao) {
+    public TimeService(TimeDao timeDao, ReservationDao reservationDao) {
         this.timeDao = timeDao;
+        this.reservationDao = reservationDao;
     }
 
     public List<Time> findAll() {
@@ -23,13 +28,13 @@ public class TimeService {
 
     public Time findById(Long id) {
         return timeDao.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 시간입니다."));
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 시간입니다."));
     }
 
     @Transactional
     public Time create(TimeRequestDto timeRequest) {
         if (timeDao.existsByStartAt(timeRequest.startAt())) {
-            throw new IllegalArgumentException("이미 존재하는 시간 입니다.");
+            throw new ConflictException("이미 존재하는 시간 입니다.");
         }
         Time time = new Time(timeRequest.startAt());
         return timeDao.insert(time);
@@ -38,8 +43,10 @@ public class TimeService {
     @Transactional
     public void delete(Long id) {
         Time time = timeDao.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 시간입니다."));
-
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 시간입니다."));
+        if (reservationDao.existsByTimeId(time.getId())) {
+            throw new ConflictException("예약이 존재하여 시간을 삭제할 수 없습니다.");
+        }
         timeDao.delete(time.getId());
     }
 }
